+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - `helmet.hidePoweredBy` is no longer a separate package. This should have no effect on end users.
 - `helmet.noSniff` is no longer a separate package. This should have no effect on end users.
+- `helmet.permittedCrossDomainPolicies` is no longer a separate package. This should have no effect on end users.
 
 ## 3.23.3 - 2020-06-26